Change order of entries in state array (#12)
* Change order of entries in state array

This change reorders the state array to make a little more sense, an application that displays the state array in order was previously very difficult to read. This groups entries by system status, trouble status, alarm status, and arming type. When the system is armed or disarmed after the initial state is loaded, the additional entry 'armed' is added to the end of the array, and the grouping is maintained. I kept 'alpha' before 'ac_present' so it did not break up the booleans 1 entry in

* Fixes and additions to state array

Renames armed_bypass to bypass since this is true even when not armed. Breaking change. Also adds an entry for the internal partition state for Honeywell panels to report the true status, as HomeAssistant does not currently support the same arming modes as the panel, namely INSTANT