Fix last_fault attribute
This attribute used to be based on the result of a zone timer dump but since they are no longer used, the value is now updated based on zone update events (DSC) and virtual keypad updates (Honeywell).  The value is now also stored as an epoch timestamp rather than "seconds since last fault".